add checking owner before delete and update

+++ backend/src/tickets/tickets.controller.ts
@@ -65,11 +65,14 @@ export class TicketsController {
   @UseGuards(JwtGuard)
   @Post()
   async create(
-    @Req() req: Request,
+    @Req() req: any,
     @Body() createPostDto: Tickets,
     @Res() res: Response,
   ) {
     try {
+      console.log('method createticket payload');
+      console.log(req.user);
+      createPostDto.userId = req.user.id;
       await this.ticketRepository.create(createPostDto);
       res.status(HttpStatus.CREATED).send({
         message: 'ticket is created',
@@ -82,43 +85,58 @@ export class TicketsController {
   @UseGuards(JwtGuard)
   @Patch(':id')
   async updated(
+    @Req() req: any,
     @Param('id') id: string,
     @Body() updatePostDto: Tickets,
     @Res() res: Response,
   ) {
     try {
-      const updateTicketCount = await this.ticketRepository.edit(
-        id,
-        updatePostDto,
-      );
-      if (updateTicketCount === 0) {
-        res
-          .status(HttpStatus.NOT_FOUND)
-          .send({ message: 'No tickets in database' });
-      } else {
-        res.status(HttpStatus.OK).send({
-          message: 'ticket ' + id + ' is updated',
-        });
+      const owner = await this.ticketRepository.find(id);
+      if (owner.userId === req.user.id) {
+        updatePostDto.userId = req.user.id;
+        const updateTicketCount = await this.ticketRepository.edit(
+          id,
+          updatePostDto,
+        );
+        if (updateTicketCount === 0) {
+          res
+            .status(HttpStatus.NOT_FOUND)
+            .send({ message: 'No tickets in database' });
+        } else {
+          res.status(HttpStatus.OK).send({
+            message: 'ticket ' + id + ' is updated',
+          });
+        }
       }
+      res.status(HttpStatus.FORBIDDEN).send({
+        message: 'FORBIDDEN',
+      });
     } catch (error) {
+      console.log(error);
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).send('Server Error');
     }
   }
 
   @UseGuards(JwtGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string, @Res() res: Response) {
+  async remove(@Req() req: any, @Param('id') id: string, @Res() res: Response) {
     try {
-      const deleteTicketCount = await this.ticketRepository.delete(id);
-      if (deleteTicketCount === 0) {
-        res
-          .status(HttpStatus.NOT_FOUND)
-          .send({ message: 'No tickets in database' });
-      } else {
-        res
-          .status(HttpStatus.OK)
-          .send({ message: 'ticket id ' + id + ' deleted successfully' });
+      const owner = await this.ticketRepository.find(id);
+      if (owner.userId === req.user.id) {
+        const deleteTicketCount = await this.ticketRepository.delete(id);
+        if (deleteTicketCount === 0) {
+          res
+            .status(HttpStatus.NOT_FOUND)
+            .send({ message: 'No tickets in database' });
+        } else {
+          res
+            .status(HttpStatus.OK)
+            .send({ message: 'ticket id ' + id + ' deleted successfully' });
+        }
       }
+      res.status(HttpStatus.FORBIDDEN).send({
+        message: 'FORBIDDEN',
+      });
     } catch (error) {
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).send('Server Error');
     }
